refactor: move IODirection into its own file
